fix(search): search acknowledges there are breads
The original test that tested against no mention of the word "bread" or
"crumb" yielded no results.  However, we now talk about them in our
pages, so they have results.  Right now we ensure that they do exist
there.  Another test might be to visit those that are returned and
verify that the text actually exists, but that could be cumbersome, so
for now we just know that there should be 2 returned and test against
that.