refactor(tabKey): create better way to handle tab
Use currying to preset variables to be able to use pressing tab key
repeatedly more easily.

pressTabKeyRepeatedly(page: Page, times: number | TabCounts)

TabCounts is a multidimensional array storing the number of tabs for a
browser and platform.  For instance:

{
    chromium: {
        mobile: 5,
        desktop: 19,
    },
    safari: {
        mobile: 19,
        desktop: 8,
    },
    default: 5,
}

In the above case, it will attempt to use the number of tabs based on
what it found in the array, and if it does not exist, it will default to
5.  In the even you forget the default, it'll be 0.