fix(YSP-315): look for button based on regex
On my system, firefox auto plays the video for the grand hero.  On
Safari, Chromium, and Mobile Safari, it does not.  As such, the button
identifier actually changes based on the state of the video playing.
This uses a regex in order to capture either since we are only concerned
with whether the button exists and is showing, not what it says
necessarily yet.