YSP-315: E2E Testing V2 (#3)
* feat(playwright): specify YALESITES_URL override

* fix(search): search acknowledges there are breads

The original test that tested against no mention of the word "bread" or
"crumb" yielded no results.  However, we now talk about them in our
pages, so they have results.  Right now we ensure that they do exist
there.  Another test might be to visit those that are returned and
verify that the text actually exists, but that could be cumbersome, so
for now we just know that there should be 2 returned and test against
that.

* fix(search): search test uses unused search

It turns out we like to use "blah", so changed this to be something
rediculous that would never be on one of our pages.

* feat(accordion): add accordion tests

* feat(accordion): test accordion functionality

* feat(accordion): works for more browsers

Found out that our menu animation is causing tab issues in testing.  At
times it is tabbing into the menu before it animates away, causing false
positives.  Waiting the short bit seems to fix the issue, but it is
still a race condition.

* feat(snapshots): change snapshot folder

Change the snapshot folder from inside the tests folder to a snapshots folder for better separation.

* feat(action_banner): add action banner tests

* feat(button-link): add button link tests

* feat(calendar-list): add calendar list tests

* feat(callout): add callout tests

* feat(visreg): all visreg tests

This was an existing script we used to test between multiple sites.
We're including it here for now as to be able to use it and integrate it
into our overall testing here.

* feat(script): Script to compare two sites

* feat(custom-cards): tests for custom card

* feat(directory): add directory tests

* feat(divider): add divider tests

* feat(images): add base images

* feat(embed): add embed tests

* feat(event): add event tests

* feat(gallery): add gallery tests

* feat(grand-hero): add grand-hero tests

* feat(image): add image tests

* feat(media-grid): add media-grid tests

* feat(post-feed): add post-feed tests

* feat(pre-built-form): add pre-built-form tests

* feat(profile): add profile tests

* feat(profile): actually test clipboard copy

Use a mock clipboard in order to determine if the copy button actually
copies text into the clipboard.  Thank you to krachtstefan for their
mockClipboardAPI implementation:
https://github.com/microsoft/playwright/issues/13037#issuecomment-1740643562z

* feat(quick-links): add quick-links tests

* feat(quote): add quote tests

* feat(spotlight-landscape): add spotlight-landscape tests

* feat(spotlight-portrait): add spotlight-portrait tests

* feat(tabs): add tab tests

* feat(text): add text tests

* feat(video): add video tests

* feat(view): add view tests minus pagination

* feat(wrapped-image): add wrapped-image tests

* feat(link-treatments): add link treatment tests

* chore(pre-build-form): rename to pre-built-form

* chore(YSP-315): add screenshot on failures

* chore(YSP-315): finish moving search

* feat(search): test search for profiles

In our last release, we found that profiles were not being indexed for
search.  This ensures that we can test that profiles are being indexed
going forward.

* feat(comparison): compare visreg on ratio

Compare success of pages based on pixel ratio over pixels.  In this last
release test, the addition of link treatments on the page caused false
positives.  This gives enough leeway to allow variations while still
identifying big issues.

* feat(npm): add shortcut scripts

You can now use:
  - Test: npm run test
  - Run tests in UI: npm run ui
  - Run tests in debug mode: npm run debug

If you're making tests, and you prefer command line reporting, you can
use: npm run dev

The other scripts will use HTML as their reporter output.

* feat(playwright): default back to html reporter

* docs(README): update with new execution examples

* feat(directory): add email test

The email in a directory is now labelled "Email", and is a link with a
mailto to their entered email address.  This tests against that.

* feat(YSP-315): add newly released snapshots

This is done so that we can easily compare on next release.  This is a
capture of any pages that have changed since release so we can visually
regression test them next time.

* chore(YSP-315): temporarily skip keyboard tests

Until YSP-352 can be fixed, tabs will be unreliable.  For now, we're
inserting a test skip with the intention of re-enabling them once
YSP-352 is complete and work starts on YSP-378: Create E2E keyboard
tests for components.

* fix(YSP-315): fix tests after release

* fix(YSP-315): grand hero videos should autoplay

* fix(YSP-315): look for button based on regex

On my system, firefox auto plays the video for the grand hero.  On
Safari, Chromium, and Mobile Safari, it does not.  As such, the button
identifier actually changes based on the state of the video playing.
This uses a regex in order to capture either since we are only concerned
with whether the button exists and is showing, not what it says
necessarily yet.

* docs(YSP-315): add browser download instructions