fix(YSP-409): post-feed now passes
With us creating new posts as part of our tests, this caused the current
test to fail since it pushed the old entries out of the first page.

I will be working on a way to run these component tests first, but until
then, we will generalize the test to make sure there are entries over
specific entries.