feat(YSP-409): test callout
This tests the callout block.  With how this one occurs, it changes the
DOM in such a way that Playwright just does not update without a timeout
interval in between.  I hate using the timeout, but have set it to 2.5
seconds to make sure we have enough time for DOM update.